new search for kwic and knew seperated buttons for search
